DEV: use LoadMore component for discovery topics list (#32363)
Follow on from https://github.com/discourse/discourse/pull/32285:

This replaces the use of the scrolling-based load-more mixin in
DiscoveryTopicsList with the IntersectionObserver-based LoadMore
component.

I insert the LoadMore component in the parent template for more precise
control over where the sentinel element is placed - right after the List
collection.

Also lifted up the `loadMore` action to the topics glimmer component so
it can be passed directly into LoadMore, and refactored away the jquery
dependency.

This is a breaking change to the behaviour of DiscoveryTopicsList, so
it'll require updates to any other repositories depending on this
component.