FEATURE: Add `routeHistory` service, SessionStore service and make `back to forum` button go back to previous page (#32260)
1000% copied from what we have in the chat

https://github.com/discourse/discourse/blob/08d5cf01cda86f1dfd48f33487b55189342532f5/plugins/chat/assets/javascripts/discourse/services/chat-state-manager.js#L164-L172

Created a service that will return the last visited URLs.

Also created a sessionStore service.

And added it to the `back-to-forum` component for use.

Maybe we could use it in the `BackButton` component, too.

Then, following the [Rule of
Three](https://blog.codinghorror.com/rule-of-three/), update the chat to
use this service as well.