FIX: key optimized images on format
Previous to this change when we used to specify format it could get an image
in the incorrect format.

Also... allows image magick to decode svg

---

There remains a bug where the crop / resize information is not stored in optimized images
this means that sometimes when asking for a cropped image you may get a resized one.