UX: composer starts with allowPreview default false (#32589)
The rich editor is loaded asynchronously, so when `allowPreview` starts
as `true`, if we're opening the composer for the 1st time, ProseMirror
dependencies will be loaded async and while the rich editor is fully
loaded, the composer animates from having a preview to not having one,
shrinking in width.

This changes the `allowPreview` to be default `false`, so the animation
never happens – when the default composer editor is set to the Markdown
editor, it'll instantly and synchronously set the tracked property,
before any animation kicks in.