FIX: Badge not showing as disabled when it is (#32591)
In the /admin/badges view, badges are showing as enabled even when they are verifiably disabled in the database.

This is happening because the defaults (needed to render the "new badge" form) are overwriting the actual value.

Since the enabled default is only used for "new badge" form, move it to the route where the other "new badge" defaults are being set.