FIX: Purge ExtraLocalesController cache correctly for multisite
When Translation Overrides are changed, a message is sent on the `/i18n-flush` channel to ensure all running unicorns flush their caches. This message is not received in a database-specific context, so running `ExtraLocalesController.clear_cache!` was only affecting the default site in a multisite cluster. This commit updates it to flush caches for all sites.