DEV: revert long press changes
These changes have been made for playwright as it was hard to test a long press event on playwright given `click` won't trigger `touchstart` event with `isMobile:true` and `hasTouch:true`. You have to use `tap`, but you don't have the `delay` option on tap, so you can't make it a long tap.

Sadly this code is apparently not working correctly on Android 15. This commit will revert the modifier to what it was before and is relying on native JS to trigger the fake long press.