DEV: introduces PLAYWRIGHT_NO_VIEWPORT
The way playwright renders headed browser is different than selenium, and depending on the size of your screen you could not be able to see part of the app. To avoid this issue you can now use `PLAYWRIGHT_NO_VIEWPORT=1` which will default to a smaller viewport.

Note that to ensure we don't fail randomly on CI, the viewport on CI will always be 1400x1400, so if you develop with `PLAYWRIGHT_NO_VIEWPORT=1` you might experience failures on CI.