DEV: clean up review queue button classes (#32685)
These reject/approve buttons were getting `btn-default` classes rather
than `btn-danger` or `btn-success`, which meant the styles for the
colors had to be applied separately to `.approve-post` and
`.reject-post`. This also meant the styles needed an additional override
in themes.

Moving to our common button classes helps centralize the styles and
avoids a manual override being needed in themes.

Button appearance is the same before/after 


![image](https://github.com/user-attachments/assets/941514f6-90c4-477c-b2e2-a02b2f296acc)