DEV: revert admin users list change
This change was unfortunately introduced in b6aad28#diff-5c7ecbec34de2166b60c5381df93a4e8438a2dbb6dcb8af0203cff4462fbc64b to prevent a failing spec due to the lack of last message bus id.

This commit reverts the change associated to it which was causing the JSON response to be an object instead of an array.