DEV: Make outletArgs available as regular arguments in PluginOutlet
Historically, plugin outlet arguments had to be accessed like `@outletArgs.foo` in templates, or `this.args.outletArgs.foo` in javascript. Following this commit, outletArgs will be passed to connector components at the top level, so they can be accessed like `@foo` in templates or `this.args.foo` in JS.

`@outletArgs` remains available for compatibility, and there are no plans to remove it.

For custom classic component connectors, these new arguments may clash with functions/fields defined on the component class. This rare case will be automatically detected, protected against, and a deprecation will be shown.