DEV: Silence scss mixed-decls for themes/plugins
Fixing this everywhere will be a very large undertaking, so we're deferring it until there is a concrete timeline for it becoming an error in sass.

In the meantime, we'll be adding a stylelint rule to enforce ordering of declarations/nested-blocks/mixins, so that will go some way towards avoiding this deprecation.