FIX: Allow existing users to accept invites that add them to a group. (#32762)
Check to see if the user needs to be added to any groups before redirecting. To avoid introducing a CSRF issue, force the user through the normal invite acceptance flow, ensuring there's user interaction before being added to any groups.

This change also tweaks the behaviour of 2e10fe9 to not follow redirects of expired invitations.

Co-authored-by: Krzysztof Kotlarek <kotlarek.krzysztof@gmail.com>