FIX: check if sidebar hidden and remove scrollLock when hiding hamburger menu (#32775)
Reported here:
https://meta.discourse.org/t/dashboard-freezes-until-sidebar-is-opened/366402

This is a follow-up to
https://github.com/discourse/discourse/pull/32651, when navigating from
the homepage to admin the hamburger gets force-hidden because we switch
from the dropdown to the sidebar nav but the scrollLock was
unintentionally left behind.

This makes sure we're only messing with hamburger state when the sidebar
is hidden (with ` {{hideApplicationSidebar}}`) and also removes the
scroll lock when the hamburger is force-hidden.