DEV: removes unnecessary lines
Technically we want to follow DDAU but most of our usage in the code base are just mutating `@checked`. This will eventually get converted to form-kit but in the meantime we can just remove this bit of code as it's doing two times the same thing and you could even end up in the situation where they both fight for what is the correct value.