DEV: allows to pass `@currentWhen` to `NavItem`
Some cases are more complex than the default behavior of `this.router.isActive(this.args.route)`, in this case you can give a function as `@currentWhen` to your `NavItem` component.

Example:

```
@action
isItCurrent(route) {
  return true;
}

<template>
  <NavItem @i18nLabel="test" @currentWhen={{this.isItCurrent}} />
</template>
```

No test as I can't write a component test for this as it relies on the router.