DEV: Port install theme button modal to new themes page (#32834)
We recently added a new themes and components page. We didn't port over the code that enables the Install theme button on Meta. It works by looking for certain query parameters and opening a special version of the install modal.

This PR is a lift-and-shift of that code from the old themes page.