FIX: Fix category title link to update when category changes, add tests (#32840)
The component was importing `@ember/component` but not marking the
category arg as tracked. Changed to `@glimmer/component` so the category
arg is tracked, and cleaned up the rest of the component to match.

Related to [this issue on
meta](https://meta.discourse.org/t/category-headers-missing-the-titles/361954).

Apparently a few customers with themes using the CategoryTitleLink
component got hit with the category title not updating correctly when
the category changed.