DEV: add new application route behavior transformers
This adds two new BehaviorTransformers

- application-route-loading
- application-route-will-transition

that might be useful for plugins interested in allow or deny listing some routes.

Both are needed because of how Ember routing works differently when the application is first loaded vs when it's already loaded and you're navigating between routes.

Also minor-bumped the plugin api version and alphabetically sorted the list of value transformers.

Internal ref - t/141359