FIX: allows zooming of images in chat
Due to one of the side effects of `app/assets/javascripts/discourse/app/lib/body-scroll-lock.js` we are disabling touch actions on body which was impacting zooming of images in chat.

Im not sure we should keep this touch-action hack  in `app/assets/javascripts/discourse/app/lib/body-scroll-lock.js` but for now the simplest change is to just disable it when a chat image is lightboxed.