UX: fix search results unreachable bottom results and more button (#32883)
Meta report:
https://meta.discourse.org/t/search-in-header-gives-too-deep-results/362973

The general problem of our `menu-panel` is that it had no proper
max-height set, which mean you have to scroll the page instead of the
panel to see bottom results. In the new header-search setup however,
scrolling the page makes the panel disappear, effectively locking you
out.

This commit adds a best-estimate max-height on the parent level
`menu-panel`, dependent on where the search is triggered from (header VS
welcome banner)

| Old | New |
|--------|--------|
| ![CleanShot 2025-05-23 at 14 58
07@2x](https://github.com/user-attachments/assets/4a96d518-13d0-4f97-b472-5d1f56bbaa1a)
| ![CleanShot 2025-05-23 at 14 59
30@2x](https://github.com/user-attachments/assets/9c8546d2-a780-49b3-80d0-95eab81eff94)
|