FIX: resolve discard draft keep editing (#32667)
Fixes a bug where clicking keep editing in the discard modal dialog
would result in the composer auto save being disabled. Because of this
updates to drafts after this point can easily be lost.

The issue was that the `cancelComposer` was not resolving the promise,
and therefore not setting `skipAutoSave` to false in the final step.

This change also includes another fix for an issue that was revealed
where the discard draft dialog would continually open again after we
close it. This would happen when editing a draft while viewing a
different topic. It wasn't noticed before as the `keep editing` button
would result in the code hanging at the `await
this.cancelComposer(opts)` stage and not opening the composer again for
the new topic reply.