FIX: Also show the toggle localization button on mobile (#32912)
In https://github.com/discourse/discourse/pull/32618, the button was
added to a place which does not show up on mobile.

When the feature was still in discourse-translator
(https://github.com/discourse/discourse-translator/pull/205), it was
added to the plugin outlet which would allow it to show up on both
mobile and desktop. This commit moves the button to be a neighbour of the
plugin outlet.