DEV: Added the ability to use users' names in group mention notifications and mentions shown in emails (#32848)
**Description**
This is part of a series of changes to allow customers to dsiplay users'
names instead of the users' username.

When a user belongs to a group that has been mentioned by another user.
It shows the name of the user that mentioned the group.

**Before**

![imagen](https://github.com/user-attachments/assets/b62224fb-9b69-4603-be00-e7aa61d9b33c)

**After**

![imagen](https://github.com/user-attachments/assets/8495cb63-6530-4d86-a51c-f0510d48f6c7)

When a email is sent to the user when mentioned in a post 

**Before**


![imagen](https://github.com/user-attachments/assets/94e674da-085a-41cb-8145-ba6fbe3636ce)

**After**

![imagen](https://github.com/user-attachments/assets/490cb365-bf85-4745-93b9-e47048b2f02e)