PERF: Remove `<details>` polyfill
When we first introduced the `discourse-details` plugin, the `<details>` element was not supported in all browsers, so this `display: none`-based polyfill had to be used.

Nowadays, there is no need for this. And in fact, the `details[open] > summary:first-of-type ~ *` selector is showing as our most expensive CSS selector during repaints.

Removing this `display: none` also means that browser find-in-page will now be able to correctly find content inside `<details>` in discourse.