FIX: Add @enabled arg to <LoadMore /> component
This commit adds an @enabled boolean to `<LoadMore />` to allow for
cases where the element is visible in the viewport but you don't want to
allow the `loadMore` behaviour. A use case for this is when our
controllers return some `canLoadMore` boolean. There is no use
attempting to load more from the server in this case, there will be
nothing else.

I also moved the CSS styles for the `<LoadMore />` component into
a new file.

This also fixes a bug in the users list shown here
https://meta.discourse.org/t/users-list-only-partial/368793