FIX: Moment locale loading in type=module
The moment locale files expect `this.` to be the window object. In a type=module, `this` is `undefined`. This commit wraps the moment-timezone definitions in an IIFE to resolve that.

Also adds a system spec to prevent future moment-timezone regressions.

Followup to a2b0c193dff40278c0a93995cc9b97b05dc8abb2